+++ src/components/primitives/index.ts
@@ -4,6 +4,7 @@ export * from "./date";
 export * from "./multiline";
 export * from "./number";
 export * from "./password";
+export * from "./radioGroup";
 export * from "./slider";
 export * from "./switch";
 export * from "./text";